develop: USING int64_t INSTEAD OF size_t WHEN DEALING WITH MEMORY-USAGE. ADVANTAGES: POSSIBILITY TO HANDLE “NEGATIVE MEMORY-USAGE” (USEFUL FOR “BALANCING”…) CONS: NONE, SINCE WE’RE NOT SUPPORTING DEVICES WITH A MEMORY-CAPACITY GREATER THAN 2~63 ;)

+++ src/chronotext/font/zf/ActualFont.cpp
@@ -264,9 +264,9 @@ namespace chr
             }
         }
         
-        size_t ActualFont::getTextureMemoryUsage() const
+        int64_t ActualFont::getTextureMemoryUsage() const
         {
-            size_t total = 0;
+            int64_t total = 0;
             
             for (auto &texture : textures)
             {
